Merged PR 92051: Add "connect to existing" functionality
Right now this is only useful if you've launched a FlexLogger application through the Python API - then you can get the server_port you need to connect to it.  We need to explore how else we expect users to want to be able to do this.

I also did a bit of cleanup I'll flag below, including adding docstrings to methods in Application.