Merged PR 93908: Add a few more tests
This adds Python versions (where appropriate) of the tests added in !93381, as well as a few others around disconnecting/closing things.

One problem it found is that if you open a project, close the application, then try to use the project, a `ValueError` will be raised (instead of an `RpcError` like happens in most other scenarios), so I modified the existing `except` clauses to handle that.  That test now passes. (as do all of them now)