Merged PR 94291: Change namespace to flexlogger.automation
After a short discussion we decided to change the namespace from `flexlogger` to `flexlogger.automation` to provide flexibility in the future. Most of these changes are mechanical, I've flagged the other ones.

I did a full run of tests and both examples to validate this, as well as building the docs.