Merged PR 99155: reduce Application.launch() timeout to 40 seconds
We got some feedback from a usability session that the default timeout to Application.launch() is too long.  (we hit this timeout if you call Application.launch() while another instance of FlexLogger is already running)

On my beefy machine, a debug build launches in ~7 seconds, so reducing the default from 60 to 40 seconds seemed reasonable.