Merged PR 99369: Basic example potpourri
Added more basic examples to cover most (but not all) of FlexLogger's entry points. The plan is to update the documentation to use these basic examples (I will do this as part of a later PR).

We are making this change based on what we observed users doing during testing: most users did not run the examples, but would hope to find code snippets they could copy and paste into their scripts when looking at the documentation.