Merged PR 107970: Show friendly error text in FlexLogger errors
We were capturing the more detailed error message correctly, but not displaying it in `repr()` (which is what gets showed in tracebacks)

I tried to calculate the more detailed error message and just pass that to the superclass constructor, but `__cause__` isn't set up then.

I added a test which used to fail and now passes, and also fixed a formatting error `black` was complaining about.

Here's what it looks like now:
![image.png](https://ni.visualstudio.com/8c36cb1d-ece7-4ec9-9c6b-409d081af0e8/_apis/git/repositories/02b121b3-366a-4aff-8bdd-b2e04aece793/pullRequests/107970/attachments/image.png)

Related work items: #1171502