Merged PR 114667: Merge from GitHub
This is a merge from the latest GitHub commit (a8105d52).  Since we haven't made any changes in this repo since we published to GitHub I did a merge with `--allow-unrelated-histories` and took all the GitHub changes in the ensuing conflicts.

I think this will make it easier to merge changes from this repo to GitHub the next time we release, although undoubtedly that will take some trial and error.