Merged PR 293985: Update version to 0.1.2 prior to merging the changes with github
Now that FlexLogger is released and available for download, follow the steps to update the public github repo to make available to users feature #1893677.

The instructions at [this page](https://dev.azure.com/ni/DevCentral/_wiki/wikis/AppCentral.wiki/17373/Managing-the-private-and-public-Git-repos) to release from the private repo to the public repo call for updating the version in setup.py.

Note that I did not find a VERSION file to update.

Also removed nptdms from the requirements as I didn't properly revert that file when I submitted my previous PR. It now matches main before the commit adding support for the disable channel functionality.