Merged PR 554388: US 2499623: Add get/set trigger settings functions, examples and tests to the FlexLogger Python API
# Justification
Intel requested the addition of new functions to the Python API to get and set the trigger settings in the logging specification.

# Implementation
Creation of the following new Python client methods to get and set the trigger settings:
- get_start_trigger_settings
- get_stop_trigger_settings
- set_test_start_trigger_settings
- set_value_change_start_trigger_settings
- set_time_start_trigger_settings
- set_test_stop_trigger_settings
- set_value_change_stop_trigger_settings
- set_time_stop_trigger_settings
- is_retriggering_enabled
- set_retriggering

Corresponding PR on the server side:
!551029

# Testing
Added new automated tests and examples for all the new methods.

Related work items: #2499623