Merged PR 563812: Fix bugs 2518733, 2522569, 2522570, 2522577 and 2522589
#Description

This PR addresses the following bugs:
#2518733
#2522569
#2522570
#2522577
#2522589

#Implementation
#2518733 was addressed by documenting the start trigger settings return value when the start trigger is set to absolute time. I also added code that converts the UTC ISO 8601 returned by the server to a datetime Python object.

#2522569
#2522570
and #2522577
were addressed on the server side, but I added tests for them on the Python client side as well.

For #2522589
I modified the `ValueChangeCondition` class to rename the `value` property `threshold` and added a `min_value` property.
Propagated these changes in all the places where we were setting `ValueChangeCondition` as well as in the function that converted the JSON string received by the server to a `ValueChangeCondition`.

#Testing
- Modified the examples.
- Fixed the broken tests.
- Added negative tests for setting invalue values for the `ValueChangeCondition`.
- Verified all tests pass.

Related work items: #2518733, #2522569, #2522570, #2522577, #2522589