Merged PR 565937: Bug 2529112: Trigger example python scripts don't work when run from console
# Justification
The project and FlexLogger was closing before waiting for the user input.
This was because the code waiting for the keystroke was not inside the have been inside the with Application.launch() as app statement.

# Implementation
Indent the code waiting for a key input and closing the project inside the `with Application.launch() as app:` statement.

# Testing
Examples now run fine.

Related work items: #2529112