Merged PR 713793: [PythonAPI] Get and remove entries from the data files pane through automation
# Justification
Customers have requested an automation method to clear the entries in the data files pane.
We also want to provide a method to obtain the entries in the data files pane.

I have added the methods to the logging specification document because they are somewhat related to the log file paths that you specify there. I can also argue adding the methods to the test session or the project.
**Please let me know if you think the methods fit better elsewhere.**

See also the corresponding changes to the automation server and LabVIEW APIs:
!713791
!713792

# Implementation
Add two methods to the automation API to:
- remove all entries in the data files pane (and optionally delete them from disk)
- get all entries in the data files pane: their paths on disk, their file creation times, and if they are backup files.

# Testing
Automated tests and manual testing through LabVIEW and Python APIs.