Merged PR 729424: Add error code property to exceptions in the Python automation API
# Justification
We have added error codes to the error messages of the FlexLogger automation API. We had to embed the error codes in the error messages. Without changes, the error codes appear in the error message:

```
Enter the name of the test property to set the value of:
Enter the test property value: 42
Failed to set test property. Additional error details: StatusCode="InvalidArgument", Detail="[-143353] The test property name must be specified when setting a test property."
Test property set. Press Enter to close the project...
```

With these changes, the error code is removed from the message and stored in a separate property on the exception:

```
Enter the name of the test property to set the value of:
Enter the test property value: 42
Failed to set test property. Additional error details: StatusCode="InvalidArgument", Detail="The test property name must be specified when setting a test property."
Test property set. Press Enter to close the project...
```

See also !730219

# Implementation
- Remove error code from error message.
- Extract error code from error message and provide value through new property on the exception.

# Testing
Manual testing with `examples/Basic/set_test_property.py`: