Merged PR 826233: Try to call the new Initialize() method when connecting to FlexLogger
# Justification

#2846519

# Implementation
Call the new Initialize method in FlexLogger.
To support older versions of FlexLogger, ignore errors if the new initialize method isn't supported.

# Testing
Manual testing:

**Old FlexLogger:**
I temporarily added a print() when ignoring the unsupported error to make sure that it was handled correctly when I run against an older FlexLogger that doesn't know about the Initialize message:
![image.png](https://dev.azure.com/ni/8c36cb1d-ece7-4ec9-9c6b-409d081af0e8/_apis/git/repositories/02b121b3-366a-4aff-8bdd-b2e04aece793/pullRequests/826233/attachments/image.png)

**Newer FlexLogger:**
Used ETW Viewer to verify that we see the Initialize action with ClientType=Python
![image (2).png](https://dev.azure.com/ni/8c36cb1d-ece7-4ec9-9c6b-409d081af0e8/_apis/git/repositories/02b121b3-366a-4aff-8bdd-b2e04aece793/pullRequests/826233/attachments/image%20%282%29.png)